v0.1.0 - Notifications
As you can see above I'm version tagging now. The webapp is useable. you can use it as a pretty reliable timer. It will also notify you when stages are changed.

Moving forward 0.X.0 is for when ANY feature is added/reworked that could even be simple sound effects and 0.0.X is for when you fix something. If you decide to save at that point.